Rename: paginationSSR common에 위치
