+++ src/components/common/PaginationSSR/PaginationSSR.module.scss
